2017-02-02  Catherine Moore  <clm@codesourcery.com>
	Merge from /csl/newlib/branches/2016.05:

	2014-12-11  Jonathan Roelofs  <jonathan@codesourcery.com>

	* libc/include/stdio.h: Fix type signature of feof and ferror in c++.

+++ newlib/libc/include/stdio.h
@@ -756,8 +756,8 @@ _ELIDABLE_INLINE int __sputc_r(struct _reent *_ptr, int _c, FILE *_p) {
 #define	__sfileno(p)	((p)->_file)
 
 #ifndef _REENT_SMALL
-#define	feof(p)		__sfeof(p)
-#define	ferror(p)	__sferror(p)
+#define	feof(p)		((int)__sfeof(p))
+#define	ferror(p)	((int)__sferror(p))
 #define	clearerr(p)	__sclearerr(p)
 
 #if __MISC_VISIBLE
